promote template var to double

+++ nya_commonmath.h
@@ -237,22 +237,22 @@ public:
 // needed to be able to swap places explicitly in multiplications/divisons
 
 template<typename T>
-inline NyaVec3Custom<T> operator*(const float a, const NyaVec3Custom<T>& b) {
+inline NyaVec3Custom<T> operator*(const double a, const NyaVec3Custom<T>& b) {
 	return b * a;
 }
 
 template<typename T>
-inline NyaVec3Custom<T> operator/(const float a, const NyaVec3Custom<T>& b) {
+inline NyaVec3Custom<T> operator/(const double a, const NyaVec3Custom<T>& b) {
 	return { a / b.x, a / b.y, a / b.z };
 }
 
 template<typename T>
-inline NyaVec4Custom<T> operator*(const float a, const NyaVec4Custom<T>& b) {
+inline NyaVec4Custom<T> operator*(const double a, const NyaVec4Custom<T>& b) {
 	return b * a;
 }
 
 template<typename T>
-inline NyaVec4Custom<T> operator/(const float a, const NyaVec4Custom<T>& b) {
+inline NyaVec4Custom<T> operator/(const double a, const NyaVec4Custom<T>& b) {
 	return { a / b.x, a / b.y, a / b.z, a / b.w };
 }
